services/board_service/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy[asyncio]
asyncmy
python-dotenv
httpx
redis